move out wandb progress wrapper
